refactor: Move `parseFallbackURL` to `useAppBootstrap.functions.ts`
